backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
geoalchemy2==0.14.2
shapely==2.0.2
redis==5.0.1
celery==5.3.4
prophet==1.1.5
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
ortools==9.8.3296
openai==1.3.7
reportlab==4.0.7
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
websockets==12.0
aioredis==2.0.1
pytest==7.4.3
httpx==0.25.2
